added single point of control for loading yolox and gdrn methods
new lines 47-49 allow the user to specify a dataset with the correct camera name in the dataset and a specific scene number. The rest of the script handles changes. However, this does not allow the user to load a SO model, but those should be combined to an ensemble model and were not considered for real-time application.